requirements.txt
Babel==2.15.0
beautifulsoup4==4.12.3
billiard==4.2.0
blessed==1.20.0
boto3==1.14.63
botocore==1.17.63
cached-property==1.5.2
celery==5.4.0
cement==2.8.2
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.3
cryptography==42.0.7
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.13
django-amazon-ses==4.0.1
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.2
django-haystack==3.2.1
django-model-utils==4.5.1
django-phonenumber-field==7.3.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-reversion==5.0.12
django-sendgrid-v5==1.2.3
django-ses==4.0.0
django-simple-history==3.5.0
django-storages==1.14.3
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-xml==2.0.0
dnspython==2.6.1
docker==3.7.3
docker-compose==1.21.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docusign-esign==3.26.0
docutils==0.15.2
drf-haystack==1.8.13
elasticsearch==7.17.9
eventlet==0.36.1
exceptiongroup==1.2.1
executing==2.0.1
factory-boy==3.3.0
Faker==25.2.0
gevent==24.2.1
greenlet==3.0.3
gunicorn==22.0.0
idna==2.6
ipython==8.24.0
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==0.10.0
jsonschema==2.6.0
kombu==5.3.7
Markdown==3.6
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
nose==1.3.7
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==24.0
parso==0.8.4
pathspec==0.5.5
pexpect==4.9.0
phonenumberslite==8.13.37
pillow==10.3.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py-oauth2==0.0.10
pyasn1==0.6.0
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-http-client==3.3.7
pytz==2024.1
PyYAML==3.13
redis==5.0.4
requests==2.18.4
requests-aws4auth==1.2.3
rsa==4.5
s3transfer==0.3.7
semantic-version==2.5.0
sendgrid==6.11.0
simplejson==3.19.2
six==1.11.0
soupsieve==2.5
sqlparse==0.5.0
stack-data==0.6.3
starkbank-ecdsa==2.2.0
tabulate==0.7.5
termcolor==1.1.0
text-unidecode==1.3
texttable==0.9.1
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.22
vine==5.1.0
Wand==0.6.13
wcwidth==0.2.13
websocket-client==0.59.0
zope.event==5.0
zope.interface==6.4

# Pinned version fixes race condition around multiple refresh calls
# Will be included in django-oauth-toolkit (which was not released as of this time)
# git+git://github.com/jazzband/django-oauth-toolkit.git@fed914c993eb51d7ec0d65798139da19082cf260
